Extract column values strategy factories into separate module
